requirements.txt

# Core dependencies
streamlit>=1.30.0
openai>=1.0.0

# ML & Clustering
scikit-learn>=1.3.0
scipy>=1.11.0
umap-learn>=0.5.5

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Visualization
plotly>=5.18.0
pandas>=2.1.0

# Utils
python-dateutil>=2.8.0